Move out ChartState as separate component
