+++ src/components/expanded-state/ChartExpandedState.js
@@ -1,13 +1,7 @@
-import { debounce, find } from 'lodash';
-import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
+import { find } from 'lodash';
+import React, { useRef } from 'react';
 import styled from 'styled-components';
-import {
-  useChartData,
-  useChartDataLabels,
-  useColorForAsset,
-  useUniswapAssetsInWallet,
-} from '../../hooks';
-
+import { useColorForAsset, useUniswapAssetsInWallet } from '../../hooks';
 import {
   BuyActionButton,
   SendActionButton,
@@ -22,19 +16,11 @@ import {
   TokenInfoRow,
   TokenInfoSection,
 } from '../token-info';
-import Chart from '../value-chart/Chart';
-import {
-  ChartPathProvider,
-  monotoneCubicInterpolation,
-} from '@rainbow-me/animated-charts';
+import ChartState from './chart/ChartState';
 import AssetInputTypes from '@rainbow-me/helpers/assetInputTypes';
 
-import { useNavigation } from '@rainbow-me/navigation';
-
-import { ModalContext } from 'react-native-cool-modals/NativeStackView';
-
 const heightWithChart = android ? 630 : 606;
-const heightWithNoChart = 309;
+const heightWithoutChart = 309;
 
 const ActionRowAndroid = styled.View`
   flex-direction: row;
@@ -46,95 +32,11 @@ const ActionRowAndroid = styled.View`
 
 const ActionRow = android ? ActionRowAndroid : SheetActionButtonRow;
 
-const traverseData = (prev, data) => {
-  if (!data || data.length === 0) {
-    return prev;
-  }
-  const filtered = data.filter(({ y }) => y);
-  if (
-    filtered[0].y === prev?.nativePoints[0]?.y &&
-    filtered[0].x === prev?.nativePoints[0]?.x
-  ) {
-    return prev;
-  }
-  const points = monotoneCubicInterpolation({
-    data: filtered,
-    includeExtremes: true,
-    range: 100,
-  });
-  return {
-    nativePoints: filtered,
-    points,
-  };
-};
-
-function useJumpingForm(isLong) {
-  const { setOptions } = useNavigation();
-
-  const { jumpToShort, jumpToLong } = useContext(ModalContext) || {};
-
-  useEffect(() => {
-    if (!isLong) {
-      setOptions({
-        isShortFormEnabled: true,
-      });
-      setImmediate(() => {
-        jumpToShort?.();
-        setOptions({
-          isShortFormEnabled: false,
-          longFormHeight: heightWithNoChart,
-        });
-      });
-    } else {
-      setOptions({
-        longFormHeight: heightWithChart,
-      });
-      setImmediate(jumpToLong);
-    }
-  }, [isLong, setOptions, jumpToShort, jumpToLong]);
-}
-
 export const ChartExpandedStateSheetHeight =
   heightWithChart + (android ? 40 : 0);
 
 export default function ChartExpandedState({ asset }) {
   const color = useColorForAsset(asset);
-  const [isFetchingInitially, setIsFetchingInitially] = useState(true);
-
-  const { chart, chartType, fetchingCharts, ...chartData } = useChartData(
-    asset
-  );
-
-  const [throttledPoints, setThrottledPoints] = useState(() =>
-    traverseData({ nativePoints: [], points: [] }, chart)
-  );
-  useEffect(() => {
-    setThrottledPoints(prev => traverseData(prev, chart));
-  }, [chart]);
-
-  const initialChartDataLabels = useChartDataLabels({
-    asset,
-    chartType,
-    color,
-    points: throttledPoints?.points ?? [],
-  });
-
-  useEffect(() => {
-    if (!fetchingCharts) {
-      setIsFetchingInitially(false);
-    }
-  }, [fetchingCharts]);
-
-  // Only show the chart if we have chart data, or if chart data is still loading
-  const showChart = useMemo(
-    () =>
-      throttledPoints?.points.length > 5 ||
-      throttledPoints?.points.length > 5 ||
-      (fetchingCharts && !isFetchingInitially),
-    [fetchingCharts, isFetchingInitially, throttledPoints]
-  );
-
-  useJumpingForm(showChart);
 
   const { uniswapAssetsInWallet } = useUniswapAssetsInWallet();
   const showSwapButton = find(uniswapAssetsInWallet, [
@@ -143,25 +45,6 @@ export default function ChartExpandedState({ asset }) {
   ]);
 
   const needsEth = asset.address === 'eth' && asset.balance.amount === '0';
-
-  const [throttledData, setThrottledData] = useState({
-    nativePoints: throttledPoints.nativePoints,
-    points: throttledPoints.points,
-    smoothingStrategy: 'bezier',
-  });
-
-  const debouncedSetThrottledData = useRef(debounce(setThrottledData, 30))
-    .current;
-
-  useEffect(() => {
-    if (throttledPoints.points && !fetchingCharts) {
-      debouncedSetThrottledData({
-        nativePoints: throttledPoints.nativePoints,
-        points: throttledPoints.points,
-        smoothingStrategy: 'bezier',
-      });
-    }
-  }, [throttledPoints, fetchingCharts, debouncedSetThrottledData]);
 
   const duration = useRef(0);
 
@@ -175,20 +58,11 @@ export default function ChartExpandedState({ asset }) {
       contentHeight={ChartExpandedStateSheetHeight}
       scrollEnabled={false}
     >
-      <ChartPathProvider data={throttledData}>
-        <Chart
-          {...chartData}
-          {...initialChartDataLabels}
-          asset={asset}
-          chart={chart}
-          chartType={chartType}
-          color={color}
-          fetchingCharts={fetchingCharts}
-          nativePoints={chart}
-          showChart={showChart}
-          throttledData={throttledData}
-        />
-      </ChartPathProvider>
+      <ChartState
+        asset={asset}
+        heightWithChart={heightWithChart}
+        heightWithoutChart={heightWithoutChart}
+      />
       <SheetDivider />
       <TokenInfoSection>
         <TokenInfoRow>
@@ -203,7 +77,7 @@ export default function ChartExpandedState({ asset }) {
         </TokenInfoRow>
       </TokenInfoSection>
       {needsEth ? (
-        <ActionRow key={`row${showChart}`}>
+        <ActionRow key="buyActionRow">
           <BuyActionButton
             color={color}
             radiusAndroid={24}
@@ -215,7 +89,7 @@ export default function ChartExpandedState({ asset }) {
           />
         </ActionRow>
       ) : (
-        <ActionRow key={`row${showChart}`}>
+        <ActionRow key="actionRow">
           {showSwapButton && (
             <SwapActionButton
               color={color}
